770GroupProject/requirements.txt
google-cloud-speech==0.30.0